chore: move components to seperate file
